+++ front/src/component/KakaoLoginComponent/index.js
@@ -1,0 +1,1 @@
+export { default } from './KakaoLoginComponent';
